Refactor MCP tests by behavior owner
Split aggregate descriptor and workflow tests into production-owner test
classes, and consolidate JDBC metadata loader scenarios under its owner.
