+++ mcp/features/readwrite-splitting/src/test/java/org/apache/shardingsphere/mcp/feature/readwritesplitting/tool/handler/PlanReadwriteSplittingRuleToolHandlerTest.java
@@ -17,17 +17,15 @@
 
 package org.apache.shardingsphere.mcp.feature.readwritesplitting.tool.handler;
 
-import org.apache.shardingsphere.mcp.api.session.MCPSessionIdentity;
 import org.apache.shardingsphere.mcp.api.payload.MCPSuccessPayload;
+import org.apache.shardingsphere.mcp.api.session.MCPSessionIdentity;
 import org.apache.shardingsphere.mcp.feature.readwritesplitting.ReadwriteSplittingFeatureDefinition;
 import org.apache.shardingsphere.mcp.feature.readwritesplitting.TestWorkflowSessionContext;
 import org.apache.shardingsphere.mcp.feature.readwritesplitting.tool.model.ReadwriteSplittingRuleWorkflowRequest;
-import org.apache.shardingsphere.mcp.feature.readwritesplitting.tool.model.ReadwriteSplittingStatusWorkflowRequest;
 import org.apache.shardingsphere.mcp.feature.readwritesplitting.tool.service.ReadwriteSplittingRuleWorkflowPlanningService;
-import org.apache.shardingsphere.mcp.feature.readwritesplitting.tool.service.ReadwriteSplittingStatusWorkflowPlanningService;
+import org.apache.shardingsphere.mcp.support.MCPFeatureRequestContext;
 import org.apache.shardingsphere.mcp.support.database.spi.MCPFeatureExecutionFacade;
 import org.apache.shardingsphere.mcp.support.database.spi.MCPFeatureQueryFacade;
-import org.apache.shardingsphere.mcp.support.MCPFeatureRequestContext;
 import org.apache.shardingsphere.mcp.support.workflow.WorkflowSessionContext;
 import org.apache.shardingsphere.mcp.support.workflow.model.ClarifiedIntent;
 import org.apache.shardingsphere.mcp.support.workflow.model.InteractionPlan;
@@ -35,7 +33,6 @@ import org.apache.shardingsphere.mcp.support.workflow.model.RuleArtifact;
 import org.apache.shardingsphere.mcp.support.workflow.model.WorkflowContextSnapshot;
 import org.apache.shardingsphere.mcp.support.workflow.model.WorkflowKind;
 import org.apache.shardingsphere.mcp.support.workflow.model.WorkflowLifecycle;
-import org.apache.shardingsphere.mcp.support.workflow.model.WorkflowRequest;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.MockedConstruction;
@@ -45,7 +42,6 @@ import java.util.Map;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
-import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
@@ -54,13 +50,13 @@ import static org.mockito.Mockito.mockConstruction;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-class ReadwriteSplittingWorkflowToolHandlersTest {
+class PlanReadwriteSplittingRuleToolHandlerTest {
     
     @Test
-    void assertHandlePlanRule() {
+    void assertHandle() {
         try (
                 MockedConstruction<ReadwriteSplittingRuleWorkflowPlanningService> mocked = mockConstruction(
-                        ReadwriteSplittingRuleWorkflowPlanningService.class, (mock, context) -> when(mock.plan(any(), any(), any())).thenReturn(createRuleSnapshot()))) {
+                        ReadwriteSplittingRuleWorkflowPlanningService.class, (mock, context) -> when(mock.plan(any(), any(), any())).thenReturn(createSnapshot()))) {
             WorkflowContextFixture fixture = createWorkflowContextFixture();
             MCPSuccessPayload actual = new PlanReadwriteSplittingRuleToolHandler().handle(fixture.requestContext, Map.of(
                     "database", "logic_db",
@@ -75,24 +71,24 @@ class ReadwriteSplittingWorkflowToolHandlersTest {
     }
     
     @Test
-    void assertHandlePlanRuleWithTopLevelLoadBalancerProperties() {
-        ReadwriteSplittingRuleWorkflowRequest actual = handlePlanRule(Map.of("database", "logic_db",
+    void assertHandleWithTopLevelLoadBalancerProperties() {
+        ReadwriteSplittingRuleWorkflowRequest actual = handle(Map.of("database", "logic_db",
                 "load_balancer_properties", Map.of("read_ds_0", "2")));
         assertThat(actual.getLoadBalancerProperties(), is(Map.of("read_ds_0", "2")));
     }
     
     @Test
-    void assertHandlePlanRuleWithStructuredIntentLoadBalancerProperties() {
-        ReadwriteSplittingRuleWorkflowRequest actual = handlePlanRule(Map.of("database", "logic_db",
+    void assertHandleWithStructuredIntentLoadBalancerProperties() {
+        ReadwriteSplittingRuleWorkflowRequest actual = handle(Map.of("database", "logic_db",
                 "structured_intent_evidence", Map.of("load_balancer_properties", Map.of("read_ds_0", "2"))));
         assertThat(actual.getLoadBalancerProperties(), is(Map.of("read_ds_0", "2")));
     }
     
     @Test
-    void assertHandlePlanRuleWithArtifacts() {
+    void assertHandleWithArtifacts() {
         try (
                 MockedConstruction<ReadwriteSplittingRuleWorkflowPlanningService> ignored = mockConstruction(
-                        ReadwriteSplittingRuleWorkflowPlanningService.class, (mock, context) -> when(mock.plan(any(), any(), any())).thenReturn(createRuleSnapshot()))) {
+                        ReadwriteSplittingRuleWorkflowPlanningService.class, (mock, context) -> when(mock.plan(any(), any(), any())).thenReturn(createSnapshot()))) {
             MCPSuccessPayload actual = new PlanReadwriteSplittingRuleToolHandler().handle(
                     createWorkflowContextFixture().requestContext, Map.of("database", "logic_db", "rule", "readwrite_ds"));
             Map<String, Object> actualPayload = actual.toPayload();
@@ -105,50 +101,10 @@ class ReadwriteSplittingWorkflowToolHandlersTest {
         }
     }
     
-    @Test
-    void assertHandlePlanStatus() {
-        try (
-                MockedConstruction<ReadwriteSplittingStatusWorkflowPlanningService> mocked = mockConstruction(
-                        ReadwriteSplittingStatusWorkflowPlanningService.class, (mock, context) -> when(mock.plan(any(), any(), any())).thenReturn(createStatusSnapshot()))) {
-            WorkflowContextFixture fixture = createWorkflowContextFixture();
-            MCPSuccessPayload actual = new PlanReadwriteSplittingStatusToolHandler().handle(fixture.requestContext, Map.of(
-                    "database", "logic_db",
-                    "target_status", "disable",
-                    "structured_intent_evidence", Map.of("rule", "readwrite_ds", "storage_unit", "read_ds_0")));
-            Map<String, Object> actualPayload = actual.toPayload();
-            assertThat(actualPayload.get("workflow_kind"), is("readwrite.status"));
-            List<?> actualResourcesToRead = (List<?>) actualPayload.get("resources_to_read");
-            assertThat(findResourceKind(actualResourcesToRead, "shardingsphere://features/readwrite-splitting/databases/logic_db/status"), is("rule"));
-            Map<?, ?> actualIntentInference = (Map<?, ?>) actualPayload.get("intent_inference");
-            assertFalse(actualIntentInference.containsKey("operation_type"));
-            assertThat(actualIntentInference.get("target_status"), is("disable"));
-            Map<?, ?> actualArgumentProvenance = (Map<?, ?>) actualPayload.get("argument_provenance");
-            assertFalse(actualArgumentProvenance.containsKey("operation_type"));
-            assertThat(actualArgumentProvenance.get("target_status"), is("user_provided"));
-            Map<?, ?> actualDistSQLArtifact = (Map<?, ?>) ((List<?>) actualPayload.get("distsql_artifacts")).getFirst();
-            assertFalse(actualDistSQLArtifact.containsKey("operation_type"));
-            assertThat(actualDistSQLArtifact.get("target_status"), is("disable"));
-            ArgumentCaptor<ReadwriteSplittingStatusWorkflowRequest> requestCaptor = ArgumentCaptor.forClass(ReadwriteSplittingStatusWorkflowRequest.class);
-            verify(mocked.constructed().getFirst()).plan(eq(fixture.workflowSessionContext), eq(fixture.queryFacade), requestCaptor.capture());
-            assertThat(requestCaptor.getValue().getStorageUnit(), is("read_ds_0"));
-            assertThat(requestCaptor.getValue().getTargetStatus(), is("disable"));
-        }
-    }
-    
-    private WorkflowContextSnapshot createRuleSnapshot() {
-        ReadwriteSplittingRuleWorkflowRequest request = new ReadwriteSplittingRuleWorkflowRequest();
-        request.setDatabase("logic_db");
-        request.setRuleName("readwrite_ds");
-        WorkflowContextSnapshot result = createSnapshot(request, ReadwriteSplittingFeatureDefinition.RULE_WORKFLOW_KIND.getValue());
-        result.getRuleArtifacts().add(new RuleArtifact("create",
-                "CREATE READWRITE_SPLITTING RULE `readwrite_ds` (WRITE_STORAGE_UNIT=`write_ds`, READ_STORAGE_UNITS(`read_ds_0`), TRANSACTIONAL_READ_QUERY_STRATEGY='DYNAMIC')"));
-        return result;
-    }
-    
-    private ReadwriteSplittingRuleWorkflowRequest handlePlanRule(final Map<String, Object> arguments) {
+    private ReadwriteSplittingRuleWorkflowRequest handle(final Map<String, Object> arguments) {
         try (
                 MockedConstruction<ReadwriteSplittingRuleWorkflowPlanningService> mocked = mockConstruction(
-                        ReadwriteSplittingRuleWorkflowPlanningService.class, (mock, context) -> when(mock.plan(any(), any(), any())).thenReturn(createRuleSnapshot()))) {
+                        ReadwriteSplittingRuleWorkflowPlanningService.class, (mock, context) -> when(mock.plan(any(), any(), any())).thenReturn(createSnapshot()))) {
             WorkflowContextFixture fixture = createWorkflowContextFixture();
             new PlanReadwriteSplittingRuleToolHandler().handle(fixture.requestContext, arguments);
             ArgumentCaptor<ReadwriteSplittingRuleWorkflowRequest> requestCaptor = ArgumentCaptor.forClass(ReadwriteSplittingRuleWorkflowRequest.class);
@@ -157,28 +113,21 @@ class ReadwriteSplittingWorkflowToolHandlersTest {
         }
     }
     
-    private WorkflowContextSnapshot createStatusSnapshot() {
-        ReadwriteSplittingStatusWorkflowRequest request = new ReadwriteSplittingStatusWorkflowRequest();
+    private WorkflowContextSnapshot createSnapshot() {
+        ReadwriteSplittingRuleWorkflowRequest request = new ReadwriteSplittingRuleWorkflowRequest();
         request.setDatabase("logic_db");
         request.setRuleName("readwrite_ds");
-        request.setStorageUnit("read_ds_0");
-        request.setTargetStatus("disable");
-        request.setOperationType("disable");
-        WorkflowContextSnapshot result = createSnapshot(request, ReadwriteSplittingFeatureDefinition.STATUS_WORKFLOW_KIND.getValue());
-        result.getRuleArtifacts().add(new RuleArtifact("disable", "ALTER READWRITE_SPLITTING RULE `readwrite_ds` DISABLE `read_ds_0` FROM `logic_db`"));
-        return result;
-    }
-    
-    private WorkflowContextSnapshot createSnapshot(final WorkflowRequest request, final String workflowKind) {
         WorkflowContextSnapshot result = new WorkflowContextSnapshot();
         result.setPlanId("plan-1");
-        result.setWorkflowKind(WorkflowKind.valueOf(workflowKind));
+        result.setWorkflowKind(WorkflowKind.valueOf(ReadwriteSplittingFeatureDefinition.RULE_WORKFLOW_KIND.getValue()));
         result.setStatus(WorkflowLifecycle.STATUS_PLANNED);
         result.setRequest(request);
         ClarifiedIntent clarifiedIntent = new ClarifiedIntent();
         clarifiedIntent.setOperationType("create");
         result.setClarifiedIntent(clarifiedIntent);
         result.setInteractionPlan(InteractionPlan.create("plan-1", request, "Readwrite workflow plan.", List.of("review"), List.of("rules")));
+        result.getRuleArtifacts().add(new RuleArtifact("create",
+                "CREATE READWRITE_SPLITTING RULE `readwrite_ds` (WRITE_STORAGE_UNIT=`write_ds`, READ_STORAGE_UNITS(`read_ds_0`), TRANSACTIONAL_READ_QUERY_STRATEGY='DYNAMIC')"));
         return result;
     }
     
@@ -200,15 +149,14 @@ class ReadwriteSplittingWorkflowToolHandlersTest {
         MCPFeatureRequestContext result = mock(MCPFeatureRequestContext.class);
         WorkflowSessionContext workflowSessionContext = new TestWorkflowSessionContext();
         MCPFeatureQueryFacade queryFacade = mock(MCPFeatureQueryFacade.class);
-        MCPFeatureExecutionFacade executionFacade = mock(MCPFeatureExecutionFacade.class);
         when(result.getSessionIdentity()).thenReturn(new MCPSessionIdentity("session-1", "", "", Map.of()));
         when(result.getWorkflowSessionContext()).thenReturn(workflowSessionContext);
         when(result.getQueryFacade()).thenReturn(queryFacade);
-        when(result.getExecutionFacade()).thenReturn(executionFacade);
-        return new WorkflowContextFixture(result, workflowSessionContext, queryFacade, executionFacade);
+        when(result.getExecutionFacade()).thenReturn(mock(MCPFeatureExecutionFacade.class));
+        return new WorkflowContextFixture(result, workflowSessionContext, queryFacade);
     }
     
     private record WorkflowContextFixture(MCPFeatureRequestContext requestContext, WorkflowSessionContext workflowSessionContext,
-                                          MCPFeatureQueryFacade queryFacade, MCPFeatureExecutionFacade executionFacade) {
+                                          MCPFeatureQueryFacade queryFacade) {
     }
 }
