+++ mcp/features/shadow/src/test/java/org/apache/shardingsphere/mcp/feature/shadow/tool/handler/PlanShadowRuleToolHandlerTest.java
@@ -17,17 +17,15 @@
 
 package org.apache.shardingsphere.mcp.feature.shadow.tool.handler;
 
-import org.apache.shardingsphere.mcp.api.session.MCPSessionIdentity;
 import org.apache.shardingsphere.mcp.api.payload.MCPSuccessPayload;
+import org.apache.shardingsphere.mcp.api.session.MCPSessionIdentity;
 import org.apache.shardingsphere.mcp.feature.shadow.ShadowFeatureDefinition;
 import org.apache.shardingsphere.mcp.feature.shadow.TestWorkflowSessionContext;
-import org.apache.shardingsphere.mcp.feature.shadow.tool.model.ShadowAlgorithmCleanupWorkflowRequest;
-import org.apache.shardingsphere.mcp.feature.shadow.tool.model.ShadowDefaultAlgorithmWorkflowRequest;
 import org.apache.shardingsphere.mcp.feature.shadow.tool.model.ShadowRuleWorkflowRequest;
 import org.apache.shardingsphere.mcp.feature.shadow.tool.service.ShadowWorkflowPlanningService;
+import org.apache.shardingsphere.mcp.support.MCPFeatureRequestContext;
 import org.apache.shardingsphere.mcp.support.database.spi.MCPFeatureExecutionFacade;
 import org.apache.shardingsphere.mcp.support.database.spi.MCPFeatureQueryFacade;
-import org.apache.shardingsphere.mcp.support.MCPFeatureRequestContext;
 import org.apache.shardingsphere.mcp.support.workflow.WorkflowSessionContext;
 import org.apache.shardingsphere.mcp.support.workflow.model.ClarifiedIntent;
 import org.apache.shardingsphere.mcp.support.workflow.model.InteractionPlan;
@@ -35,7 +33,6 @@ import org.apache.shardingsphere.mcp.support.workflow.model.RuleArtifact;
 import org.apache.shardingsphere.mcp.support.workflow.model.WorkflowContextSnapshot;
 import org.apache.shardingsphere.mcp.support.workflow.model.WorkflowKind;
 import org.apache.shardingsphere.mcp.support.workflow.model.WorkflowLifecycle;
-import org.apache.shardingsphere.mcp.support.workflow.model.WorkflowRequest;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.MockedConstruction;
@@ -52,13 +49,13 @@ import static org.mockito.Mockito.mockConstruction;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-class ShadowWorkflowToolHandlersTest {
+class PlanShadowRuleToolHandlerTest {
     
     @Test
-    void assertHandlePlanRule() {
+    void assertHandle() {
         try (
                 MockedConstruction<ShadowWorkflowPlanningService> mocked = mockConstruction(ShadowWorkflowPlanningService.class,
-                        (mock, context) -> when(mock.planRule(any(), any(), any())).thenReturn(createSnapshot(createRuleRequest(), ShadowFeatureDefinition.RULE_WORKFLOW_KIND.getValue())))) {
+                        (mock, context) -> when(mock.planRule(any(), any(), any())).thenReturn(createSnapshot()))) {
             WorkflowContextFixture fixture = createWorkflowContextFixture();
             MCPSuccessPayload actual = new PlanShadowRuleToolHandler().handle(fixture.requestContext, Map.of(
                     "database", "logic_db",
@@ -77,45 +74,21 @@ class ShadowWorkflowToolHandlersTest {
         }
     }
     
-    @Test
-    void assertHandlePlanDefaultAlgorithm() {
-        try (
-                MockedConstruction<ShadowWorkflowPlanningService> mocked =
-                        mockConstruction(ShadowWorkflowPlanningService.class, (mock, context) -> when(mock.planDefaultAlgorithm(any(), any(), any())).thenReturn(createSnapshot(
-                                new ShadowDefaultAlgorithmWorkflowRequest(), ShadowFeatureDefinition.DEFAULT_ALGORITHM_WORKFLOW_KIND.getValue())))) {
-            WorkflowContextFixture fixture = createWorkflowContextFixture();
-            MCPSuccessPayload actual = new PlanDefaultShadowAlgorithmToolHandler()
-                    .handle(fixture.requestContext, Map.of("database", "logic_db", "algorithm_type", "SQL_HINT"));
-            ArgumentCaptor<ShadowDefaultAlgorithmWorkflowRequest> requestCaptor = ArgumentCaptor.forClass(ShadowDefaultAlgorithmWorkflowRequest.class);
-            verify(mocked.constructed().getFirst()).planDefaultAlgorithm(
-                    eq(fixture.workflowSessionContext), eq(fixture.queryFacade), requestCaptor.capture());
-            assertThat(requestCaptor.getValue().getAlgorithmType(), is("SQL_HINT"));
-            assertThat(actual.toPayload().get("workflow_kind"), is("shadow.default"));
-        }
-    }
-    
-    @Test
-    void assertHandlePlanCleanup() {
-        try (
-                MockedConstruction<ShadowWorkflowPlanningService> mocked =
-                        mockConstruction(ShadowWorkflowPlanningService.class, (mock, context) -> when(mock.planAlgorithmCleanup(any(), any(), any())).thenReturn(createSnapshot(
-                                new ShadowAlgorithmCleanupWorkflowRequest(), ShadowFeatureDefinition.ALGORITHM_CLEANUP_WORKFLOW_KIND.getValue())))) {
-            WorkflowContextFixture fixture = createWorkflowContextFixture();
-            MCPSuccessPayload actual = new PlanShadowAlgorithmCleanupToolHandler()
-                    .handle(fixture.requestContext, Map.of("database", "logic_db", "algorithm_name", "unused_algorithm"));
-            ArgumentCaptor<ShadowAlgorithmCleanupWorkflowRequest> requestCaptor = ArgumentCaptor.forClass(ShadowAlgorithmCleanupWorkflowRequest.class);
-            verify(mocked.constructed().getFirst()).planAlgorithmCleanup(
-                    eq(fixture.workflowSessionContext), eq(fixture.queryFacade), requestCaptor.capture());
-            assertThat(requestCaptor.getValue().getAlgorithmName(), is("unused_algorithm"));
-            assertThat(actual.toPayload().get("workflow_kind"), is("shadow.cleanup"));
-        }
-    }
-    
-    private ShadowRuleWorkflowRequest createRuleRequest() {
-        ShadowRuleWorkflowRequest result = new ShadowRuleWorkflowRequest();
-        result.setDatabase("logic_db");
-        result.setRuleName("shadow_rule");
-        result.setTableName("t_order");
+    private WorkflowContextSnapshot createSnapshot() {
+        ShadowRuleWorkflowRequest request = new ShadowRuleWorkflowRequest();
+        request.setDatabase("logic_db");
+        request.setRuleName("shadow_rule");
+        request.setTableName("t_order");
+        WorkflowContextSnapshot result = new WorkflowContextSnapshot();
+        result.setPlanId("plan-1");
+        result.setWorkflowKind(WorkflowKind.valueOf(ShadowFeatureDefinition.RULE_WORKFLOW_KIND.getValue()));
+        result.setStatus(WorkflowLifecycle.STATUS_PLANNED);
+        result.setRequest(request);
+        ClarifiedIntent clarifiedIntent = new ClarifiedIntent();
+        clarifiedIntent.setOperationType("create");
+        result.setClarifiedIntent(clarifiedIntent);
+        result.setInteractionPlan(InteractionPlan.create("plan-1", request, "Shadow workflow plan.", List.of("review"), List.of("rules")));
+        result.getRuleArtifacts().add(new RuleArtifact("create", "CREATE SHADOW RULE `shadow_rule`(SOURCE=`demo_ds`, SHADOW=`demo_ds_shadow`, `t_order`(TYPE(NAME='sql_hint')))"));
         return result;
     }
     
@@ -129,33 +102,18 @@ class ShadowWorkflowToolHandlersTest {
         return "";
     }
     
-    private WorkflowContextSnapshot createSnapshot(final WorkflowRequest request, final String workflowKind) {
-        WorkflowContextSnapshot result = new WorkflowContextSnapshot();
-        result.setPlanId("plan-1");
-        result.setWorkflowKind(WorkflowKind.valueOf(workflowKind));
-        result.setStatus(WorkflowLifecycle.STATUS_PLANNED);
-        result.setRequest(request);
-        ClarifiedIntent clarifiedIntent = new ClarifiedIntent();
-        clarifiedIntent.setOperationType("create");
-        result.setClarifiedIntent(clarifiedIntent);
-        result.setInteractionPlan(InteractionPlan.create("plan-1", request, "Shadow workflow plan.", List.of("review"), List.of("rules")));
-        result.getRuleArtifacts().add(new RuleArtifact("create", "CREATE SHADOW RULE `shadow_rule`(SOURCE=`demo_ds`, SHADOW=`demo_ds_shadow`, `t_order`(TYPE(NAME='sql_hint')))"));
-        return result;
-    }
-    
     private WorkflowContextFixture createWorkflowContextFixture() {
         MCPFeatureRequestContext result = mock(MCPFeatureRequestContext.class);
         WorkflowSessionContext workflowSessionContext = new TestWorkflowSessionContext();
         MCPFeatureQueryFacade queryFacade = mock(MCPFeatureQueryFacade.class);
-        MCPFeatureExecutionFacade executionFacade = mock(MCPFeatureExecutionFacade.class);
         when(result.getSessionIdentity()).thenReturn(new MCPSessionIdentity("session-1", "", "", Map.of()));
         when(result.getWorkflowSessionContext()).thenReturn(workflowSessionContext);
         when(result.getQueryFacade()).thenReturn(queryFacade);
-        when(result.getExecutionFacade()).thenReturn(executionFacade);
-        return new WorkflowContextFixture(result, workflowSessionContext, queryFacade, executionFacade);
+        when(result.getExecutionFacade()).thenReturn(mock(MCPFeatureExecutionFacade.class));
+        return new WorkflowContextFixture(result, workflowSessionContext, queryFacade);
     }
     
     private record WorkflowContextFixture(MCPFeatureRequestContext requestContext, WorkflowSessionContext workflowSessionContext,
-                                          MCPFeatureQueryFacade queryFacade, MCPFeatureExecutionFacade executionFacade) {
+                                          MCPFeatureQueryFacade queryFacade) {
     }
 }
