+++ mcp/features/sharding/src/test/java/org/apache/shardingsphere/mcp/feature/sharding/tool/handler/ShardingWorkflowToolHandlerTestFixture.java
@@ -1,0 +1,93 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shardingsphere.mcp.feature.sharding.tool.handler;
+
+import org.apache.shardingsphere.mcp.api.session.MCPSessionIdentity;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+import org.apache.shardingsphere.mcp.feature.sharding.TestWorkflowSessionContext;
+import org.apache.shardingsphere.mcp.feature.sharding.tool.model.ShardingWorkflowRequest;
+import org.apache.shardingsphere.mcp.support.MCPFeatureRequestContext;
+import org.apache.shardingsphere.mcp.support.database.spi.MCPFeatureQueryFacade;
+import org.apache.shardingsphere.mcp.support.workflow.WorkflowSessionContext;
+import org.apache.shardingsphere.mcp.support.workflow.model.ClarifiedIntent;
+import org.apache.shardingsphere.mcp.support.workflow.model.InteractionPlan;
+import org.apache.shardingsphere.mcp.support.workflow.model.RuleArtifact;
+import org.apache.shardingsphere.mcp.support.workflow.model.WorkflowContextSnapshot;
+import org.apache.shardingsphere.mcp.support.workflow.model.WorkflowKind;
+import org.apache.shardingsphere.mcp.support.workflow.model.WorkflowLifecycle;
+
+import java.util.List;
+import java.util.Map;
+
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+final class ShardingWorkflowToolHandlerTestFixture {
+    
+    static WorkflowContextSnapshot createSnapshot(final WorkflowKind workflowKind, final String sql) {
+        ShardingWorkflowRequest request = createRequest();
+        WorkflowContextSnapshot result = new WorkflowContextSnapshot();
+        result.setPlanId("plan-1");
+        result.setWorkflowKind(workflowKind);
+        result.setStatus(WorkflowLifecycle.STATUS_PLANNED);
+        result.setRequest(request);
+        ClarifiedIntent clarifiedIntent = new ClarifiedIntent();
+        clarifiedIntent.setOperationType("create");
+        result.setClarifiedIntent(clarifiedIntent);
+        result.setInteractionPlan(InteractionPlan.create("plan-1", request, "Sharding workflow plan.", List.of("review"), List.of("rules")));
+        result.getRuleArtifacts().add(new RuleArtifact("create", sql));
+        return result;
+    }
+    
+    static ShardingWorkflowRequest createRequest() {
+        ShardingWorkflowRequest result = new ShardingWorkflowRequest();
+        result.setDatabase("logic_db");
+        result.setTable("t_order");
+        return result;
+    }
+    
+    static List<String> extractResourceUris(final List<?> resourcesToRead) {
+        return resourcesToRead.stream().map(each -> (String) ((Map<?, ?>) each).get("uri")).toList();
+    }
+    
+    static String findResourceKind(final List<?> resourcesToRead, final String uri) {
+        for (Object each : resourcesToRead) {
+            Map<?, ?> resource = (Map<?, ?>) each;
+            if (uri.equals(resource.get("uri"))) {
+                return (String) resource.get("resource_kind");
+            }
+        }
+        return "";
+    }
+    
+    static WorkflowContextFixture createWorkflowContextFixture() {
+        MCPFeatureRequestContext result = mock(MCPFeatureRequestContext.class);
+        WorkflowSessionContext workflowSessionContext = new TestWorkflowSessionContext();
+        MCPFeatureQueryFacade queryFacade = mock(MCPFeatureQueryFacade.class);
+        when(result.getSessionIdentity()).thenReturn(new MCPSessionIdentity("session-1", "", "", Map.of()));
+        when(result.getWorkflowSessionContext()).thenReturn(workflowSessionContext);
+        when(result.getQueryFacade()).thenReturn(queryFacade);
+        return new WorkflowContextFixture(result, workflowSessionContext, queryFacade);
+    }
+    
+    record WorkflowContextFixture(MCPFeatureRequestContext requestContext, WorkflowSessionContext workflowSessionContext,
+                                  MCPFeatureQueryFacade queryFacade) {
+    }
+}
